copilot-c99: Adapt `mkStep` to use unique triggers. Refs #296.
The current implementation of Copilot does not allow using the same
trigger handler name multiple times, since the C99 backend produces
multiple repeated declarations, which is invalid.

A prior commit has introduced a representation of triggers with unique
IDs, allowing us to distinguish between a monitor for one condition
within a Copilot spec vs the external C function being called when the
condition becomes true.

This commit updates the generation of the `step` function to use unique
names for the triggers.