copilot-theorem: Bump upper version constraint on data-default. Refs #570.
In order to keep Copilot effectively working in the current Haskell ecosystem,
and with new versions of GHC, we want to extend the versions of dependencies
that Copilot can be installed with as much as possible to include newly
released versions.

This commit relaxes the upper bounds on the version constraints on
data-default so that Copilot can be installed with version 0.8.0.0.