copilot: Include missing dependencies in installation instructions. Refs #591.
The Linux section of the README lists GHC and Cabal as dependencies, but
more packages are needed. Specifically, in a fresh new Ubuntu as
provided via the Docker Hub, users also need to install libz-dev and
pkg-config. The programs alex and happy also need to be available.

This commit adjust the installation instructions for Linux distributions
where Copilot is not directly available as a package to include other
dependencies needed, namely libz-dev, pkg-config, alex and happy.

The text prior to the instructions is also updated to reflect that this
is no longer just the Haskell toolchain but, rather, all dependencies of
Copilot.