copilot-theorem: Reject existentially quantified propositions in What4 backend. Refs #254.
The functions in `Copilot.Theorem.What4` (e.g., `prove`) take a proposition and
attempt to prove that it is valid at all possible time steps. Previously, they
would do this regardless of whether the proposition was universally quantified
or existential quantifier, however, which was unsound. The reason that this
happened is because the functions would call `extractProp` on the proposition
being proven, causing the proposition's quantifier not to be taken into
account.

This commit removes the unsound uses of `extractProp` and instead checks if the
proposition was actually quantified universally (i.e., with a `Forall`), which
is the intended use case for `Copilot.Theorem.What4`. If the proposition was
instead quantified existentially (i.e., with an `Exists`), then a
`ProveException` will be thrown. Warnings have been added to Haddocks of the
relevant functions that can potentially throw a `ProveException`.