copilot-core: Add Show instance for Type. Refs #589.
Currently, the `Copilot.Theorem.What4.prove` function returns a list of
results, where each result contains a `SatResult` that describes whether a
property is `Valid`, `Invalid`, or `Unknown`. The `Invalid` result has the
limitation that it does not give any information about a specific
counterexample that could drive Copilot into falsifying the property, however.
This makes it challenging to interpret what the results of prove mean.

To define a companion function to `prove` that also returns counterexample
information upon a failed proof, it is convenient to be able to display `Type`
information in panic messages.

This commit derives a basic `Show` instance for `Type` so that
`copilot-theorem` can display them whenever an internal invariant is violated.