copilot-theorem: Add Show and ShowF instances for counterexamples. Refs #589.
Currently, the `Copilot.Theorem.What4.prove` function returns a list of
results, where each result contains a `SatResult` that describes whether a
property is `Valid`, `Invalid`, or `Unknown`. The `Invalid` result has the
limitation that it does not give any information about a specific
counterexample that could drive Copilot into falsifying the property, however.
This makes it challenging to interpret what the results of prove mean.

The `CounterExample`, `SatResultCex`, and `CopilotValue` data types lack `Show`
and `ShowF` instances, which makes it impractical for users to display them.

This commit adds `Show` and `ShowF` instances for all three data types so that
they can be shown.