copilot-theorem: Add Props to specProps. Refs #594.
`copilot-theorem`'s Kind2 backend takes a proposition and attempts to prove
that it is valid at all possible time steps. Currently, this happens regardless
of whether the proposition is universally or existentially quantified. This is
unsound, and it can lead to valid, existentially quantified properties being
incorrectly labeled as invalid.

This commit adds information about propositions' quantifiers to the Kind2
backend's implementation so that it may use the quantifiers when translating
the properties to Kind2. This is accomplished by updating `specProps` to map
each proposition's identifier to its corresponding `copilot-core` `Prop`.