copilot-theorem: Translate quantifiers correctly in Kind2 backend. Refs #594.
`copilot-theorem`'s Kind2 backend takes a proposition and attempts to prove
that it is valid at all possible time steps. Currently, this happens regardless
of whether the proposition is universally or existentially quantified. This is
unsound, and it can lead to valid, existentially quantified properties being
incorrectly labeled as invalid.

This commit fixes the translations of propositions to Kind2 (via Kind2's
TransSys backend) such that the translated properties respect quantifier
information. Specifically, this replaces an unsound use of the `extractProp`
function with an analysis that translates the property differently depending on
whether it is universally or existentially quantified. Because all TransSys
propositions use universal quantification, we convert an existentially
quantified property `∃x.P(x)` as `¬(∀x.¬(P(x)))` during the translation to
TransSys.