copilot-theorem: Add function to produce counterexamples for invalid properties. Refs #589.
Currently, the `Copilot.Theorem.What4.prove` function returns a list of
results, where each result contains a `SatResult` that describes whether a
property is `Valid`, `Invalid`, or `Unknown`. The `Invalid` result has the
limitation that it does not give any information about a specific
counterexample that could drive Copilot into falsifying the property, however.
This makes it challenging to interpret what the results of prove mean.

This introduces a new `proveWithCounterExample` function to
`Copilot.Theorem.What4` that mirrors the type signature of `prove`, except that
it returns a variant of `SatResult` (`SatResultCex`) where the `Invalid`
equivalent (`InvalidCex`) encodes counterexample information. `copilot-theorem`
users can then interpret the results of the counterexample in Copilot
specifications.

As part of this commit, we change the definition of the `CounterExample` data
type. This is safe to do, as `CounterExample` was completely unused prior to
this commit, nor was it exported.