copilot-theorem: Add test cases for proveWithCounterExample. Refs #589.
Currently, the `Copilot.Theorem.What4.prove` function returns a list of
results, where each result contains a `SatResult` that describes whether a
property is `Valid`, `Invalid`, or `Unknown`. The `Invalid` result has the
limitation that it does not give any information about a specific
counterexample that could drive Copilot into falsifying the property, however.
This makes it challenging to interpret what the results of prove mean.

A prior commit has introduced a `proveWithCounterExample` function, which
provides a counterexample when a property is proven invalid.

This commit updates the test suite to ensure that basic uses of
`proveWithCounterExample` work as intended.