copilot-language: Remove deprecated function Copilot.Language.Operators.Array.(.!!). Refs #599.
The function `copilot-language:Copilot.Language.Operators.Array.(.!!)` was
deprecated due to the introduction of a new, common interface to interact with
structures and arrays. The original function was deprecated in Copilot 4 and no
messages have been received requesting that it be kept in this library. Per
our internal policy of waiting 3 versions from deprecation until a public
interface declaration can be removed, this function can now be removed.

This commit removes the function
`copilot-language:Copilot.Language.Operators.Array.(.!!)`.