copilot-libraries: Remove deprecated function Copilot.Library.Utils.(!!). Refs #599.
The function `copilot-libraries:Copilot.Library.Utils.(!!)` was deprecated due
to the introduction of a new, common interface to interact with structures and
arrays. The original function was deprecated in Copilot 4 and no messages have
been received requesting that it be kept in this library. Per our internal
policy of waiting 3 versions from deprecation until a public interface
declaration can be removed, this function can now be removed.

This commit removes the function
`copilot-libraries:Copilot.Library.Utils.(!!)`, and adjusts any imports that
were hiding either the same function or equally named functions to avoid
clashes, since they would no longer occur.