copilot: Update build status URL. Refs #612.
Currently the badge for the build status in the README doesn't show up
correctly. GitHub has cached the result but it's not providing the
resulting image, so we see "Build Status" instead.

This commit updates the badge URL to the current
"https://api.travis-ci.com/Copilot-Language/copilot.svg?branch=master",
which enables the badge to show properly.