copilot-core: Remove deprecated Copilot.Core.Type.uTypeType. Refs #615.
The record field uTypeType, part of Copilot.Core.Type.UType, is never really
used within the module or in any other part of Copilot or, as far as we know,
by any user of Copilot. The record field was deprecated in Copilot 4.1 and no
messages have been received requesting that it be kept in this library. As per
our internal policy of waiting 3 versions from deprecation until a public
interface declaration can be removed, this definition can now be removed.

This commit removes the field uTypeType from the definition of
Copilot.Core.Type.UType, making the type no longer a record.