copilot-bluespec: Include in mainline copilot repo. Refs #623.
The `copilot-bluespec` package was originally developed in an external repo
(https://github.com/Copilot-Language/copilot-bluespec), but there is interest
in having all Copilot projects and libraries merged into the main
https://github.com/Copilot-Language/copilot repo for ease of management.

This commit includes the last state of the old `copilot-bluespec` repo into the
main Copilot repo as a new subdirectory, disregarding the commit history. It
also adjusts the `copilot-bluespec` changelog to make the old issues refer to
`copilot-bluespec#<ISSUE_NUM>` so that they are not interpreted as related to
the main repo.

In addition, there were several pieces of documentation throughout the repo
that referred to `copilot-c99` as being the sole Copilot backend. This commit
updates these pieces of documentation to reflect that there are now two Copilot
backends (C99 and Bluespec).

Finally, this commit updates the CI to build and test `copilot-bluespec`
alongside the other Copilot packages. Doing so requires installing `bsc`
beforehand, as `bsc` is required to compile and simulate the code that
`copilot-bluespec` generates.