copilot-libraries: Add stream analysis module. Refs #726.
There is a need to have in Copilot a module that includes stream and
spec analysis functions. Such functions have been useful in other
projects, and they could be added as part of Copilot's standard library
so that they are available to all users.

This commit adds a new module that provides functions to determine if a
stream is always true or always false.