copilot-bluespec: Include in mainline copilot repo. Refs #623.
The `copilot-bluespec` package was originally developed in an external repo
(https://github.com/Copilot-Language/copilot-bluespec), but there is interest
in having all Copilot projects and libraries merged into the main
https://github.com/Copilot-Language/copilot repo for ease of management.

This commit includes the last state of the old `copilot-bluespec` repo into the
main Copilot repo as a new subdirectory, disregarding the commit history. It
also adjusts the `copilot-bluespec` changelog to make the old issues refer to
`copilot-bluespec#<ISSUE_NUM>` so that they are not interpreted as related to
the main repo.