Explicitly list .cabal files in cabal.project. Refs #622.
The `copilot-verifier` package was originally developed in an external repo
(https://github.com/Copilot-Language/copilot-verifier), but there is interest
in having all Copilot projects and libraries merged into the main
https://github.com/Copilot-Language/copilot repo for ease of management.

For now, we wish not to add `copilot-verifier` to the `cabal.project` file, as
its version bounds are tighter than the range of GHC versions that `copilot`
currently supports. As such, this commit explicitly lists the `.cabal` files of
packages that we do want to include in the `cabal.project`, i.e., all of the
`.cabal` files except for `copilot-verifier.cabal`.