copilot-visualizer: Remove duplicate .gitignore. Refs #624.
The copilot-visualizer package was developed outside of this repo, but
there's interest in having all Copilot projects and libraries merged
into a main repo for ease of management.

A prior commit add the copilot-visualizer library as a subdirectory of
the main repository. The library included a .gitignore file, which is a
duplicate of the one already present at the top level of the repository.

This commit removes the duplicate file to have all ignored files
governed by the .gitignore file at the top level.