copilot-visualizer: Adjust issue numbers in CHANGELOG. Refs #624.
The copilot-visualizer package was originally developed in an external
repo, but there is interest in having all Copilot projects and
libraries merged into the main repo for ease of management.

A prior commit added the copilot-visualizer library as a subdirectory of
the main copilot repo. However the CHANGELOG included with that library
lists issue numbers that refer to issues in the original repo, not the
main Copilot repo, which can cause confusion.

This commit adjusts old issue numbers in the copilot-visualizer
CHANGELOG to include the package name, to make it clear that they refer
to the old repo.