copilot-core: Remove deprecated fields of Copilot.Core.Expr.UExpr. Refs #636.
The record fields uExprExpr and uExprType, part of Copilot.Core.Expr,
are never really used within the module or in any other part of Copilot
or, as far as we know, by any user of Copilot. The record fields were
deprecated in Copilot 4.2 and no messages have been received requesting
that they be kept in this library. As per our internal policy of waiting
3 versions from deprecation until a public interface declaration can be
removed, these definitions can now be removed.

This commit removes the fields uExprType and uExprExpr from the
definition of Copilot.Core.Expr.UExpr, making the type no longer a
record.