copilot: Add hello world example for FPGA. Refs #657.
We need an example that demonstrates how to work with FPGAs using Copilot. This
will help users start with copilot-bluespec, and also help validate our work.

This commit adds a hello world example that demonstrates how to control 4 LEDs
from 4 switches. The tutorial describes how to install Vivado, create the spec,
compile it (in bluespec and then verilog), and upload it to the board.