copilot-visualizer: Relax version constraint on `ogma-extra`. Refs #645.
In order to keep Copilot effectively working in the current Haskell
ecosystem, as well as new versions of GHC, we need to extend the
versions of dependencies that Copilot can be installed with. The package
`ogma-extra` has seen release 1.9, but copilot-visualizer depend on
versions prior to 1.7.

This commit relaxes the version constraint of on `ogma-extra` to allow
installation of `copilot-visualizer` with versions in the 1.9 series.