copilot-theorem: Remove unused function Copilot.Theorem.Misc.SExpr.parseSExpr. Refs #661.
The function `copilot-theorem:Copilot.Theorem.Misc.SExpr.parseSExpr` is
not used anywhere in the implementation. The module is internal and not
exposed to the user, so that function does not have and cannot have any
users.

This commit removes the function, as well as another definition in the
module only used by `parseSExpr`, and an import only used by those two
functions.