copilot-theorem: Removed unnecessary dependency on parsec. Refs #661.
The function `copilot-theorem:Copilot.Theorem.Misc.SExpr.parseSExpr` is
not used anywhere in the implementation. The module is internal and not
exposed to the user, so that function does not have and cannot have any
users.

A prior commit has removed the function, which was the only user of the
`parsec` library across `copilot-theorem`, thus rendering that
dependency no longer necessary.

This commit removes that dependency from the Cabal file.