copilot-bluespec: Update copilot-bluespec.cabal to only run tests when passed "test-bluespec" flag. Refs Copilot-Language#673.
The copilot-bluespec package requires bsc (the Bluespec compiler) to be
installed in order to run its test suite, but the builders that Hackage uses to
run uploaded packages' test suites do not have bsc installed, causing Hackage
to spuriously claim that copilot-bluespec's tests are failing.

This commit adds a cabal flag `test-bluespec` intended to enable tests that
require the Bluespec compiler (`bsc`) to run. The default behavior of `cabal
test` (with no flags) is now to not build the testsuite, preventing spurious
test failures from appearing in Hackage reports. To enable building the
testsuite, use: `cabal test -f test-bluespec`.