copilot-c99: Raise error when compiling empty array or struct. Refs #695.
Copilot allows defining empty arrays (i.e., values of type Array 0) and empty
structs (i.e., structs with no fields), but using empty arrays or structs in
C99 is undefined behavior. Moreover, copilot-c99 will often crash if you
attempt to compile a Copilot specification that uses an empty array or struct.

This commit raises an appropriate error when attempting to translate a
zero-length array or a struct with no fields to C99. Previously, Copilot would
either crash with a confusing error or silently translate the offending spec to
invalid C99.