copilot-bluespec: Flip direction of interface inputs and outputs in Bluespec. Refs #677.
The current Bluespec backend leads to Verilog code that requires manual
manipulations in order to work correctly. Specifically, Copilot externs, which
are inputs to the Copilot monitoring system, are treated as _outputs_ in
Verilog, and Copilot triggers, which can be considered outputs of the
monitoring system, are treated as _inputs_. This order of inputs and outputs is
the opposite what we would like users to work with, and of what other backends
generate (e.g., C99).

This commit updates the internals of `copilot-bluespec` to flip the order in
which generated Bluespec module interfaces declare their inputs and outputs. A
module interface now represents each Copilot extern as a method of type `<ty>
-> Action` (where `<ty>` is the type of the extern), as this type directly
translates to a Verilog input. Moreover, a module interface now represents
trigger arguments as interface methods of type `<ty>_i`, where each `<ty>_i` is
the type of each trigger argument. Each `<ty>_i` directly translates to a
Verilog output.

In addition to changing the module interface definitions, this commit also
changes the generated modules that instantiate the new interfaces. One of the
more notable changes is that all extern values are stored as a Bluespec `Wire`
in the module internals, as using `Wire`s leads to compact Verilog code without
adding any user-visible inputs or outputs. Every time a value is read from an
extern, the corresponding `Wire` must be read from, which means that several
parts of the module internals have been updated to reference these `Wire`s.