copilot-bluespec: Update Copilot.Compile.Bluespec.Compile to reflect new input/output order. Refs #677.
The current Bluespec backend leads to Verilog code that requires manual
manipulations in order to work correctly. Specifically, Copilot externs, which
are inputs to the Copilot monitoring system, are treated as _outputs_ in
Verilog, and Copilot triggers, which can be considered outputs of the
monitoring system, are treated as _inputs_. This order of inputs and outputs is
the opposite what we would like users to work with, and of what other backends
generate (e.g., C99).

A prior commit has updated the internals of `copilot-bluespec` to invert the
order in which Verilog inputs and outputs are declared. This commit adapts the
top-level code in `Copilot.Compile.Bluespec.Compile` accordingly. In
particular, the `compileWith` function now generates a `<Prefix>.bs` file that
defines two module interfaces: one interface that contains the core methods
that will comprise the generated Verilog code, and another interface that helps
automate the process of defining a `Module Empty` value that is more suitable
for software simulation purposes.

Previously, `compileWith` generated a separate `<Prefix>Ifc.bs` file that
contained nothing but the module interface in isolation so that it could be
used separately from the interface's `Module` definition. This distinction no
longer makes as much sense, as in the new design, as (1) there are now multiple
interfaces, and (2) it is unlikely that the interfaces are not particularly
useful without having the `Module` definitions to accompany them. As such, this
commit now omits generating a `<Prefix>Ifc.bs` file, instead opting to generate
the interfaces as part of `<Prefix>.bs`.