copilot-bluespec: Update test suite to match new input/output order. Refs #677.
The current Bluespec backend leads to Verilog code that requires manual
manipulations in order to work correctly. Specifically, Copilot externs, which
are inputs to the Copilot monitoring system, are treated as _outputs_ in
Verilog, and Copilot triggers, which can be considered outputs of the
monitoring system, are treated as _inputs_. This order of inputs and outputs is
the opposite what we would like users to work with, and of what other backends
generate (e.g., C99).

A prior commit has updated the internals of `copilot-bluespec` to invert the
order in which Verilog inputs and outputs are declared. This commit updates the
test suite to work with the new design. In addition to changing the code to
reflect the new order of inputs and outputs, this commit also updates the names
of some internal variables from "`input`" to "`input1`". This change is
necessary because "`input`" is a reserved keyword in Verilog, and giving a
Bluespec interface method the same name as a reserved Verilog keyword becomes
an error when the interface is used in the result type of a synthesized module.