copilot-bluespec: Update DESIGN.md to reflect new input/output order. Refs #677.
The current Bluespec backend leads to Verilog code that requires manual
manipulations in order to work correctly. Specifically, Copilot externs, which
are inputs to the Copilot monitoring system, are treated as _outputs_ in
Verilog, and Copilot triggers, which can be considered outputs of the
monitoring system, are treated as _inputs_. This order of inputs and outputs is
the opposite what we would like users to work with, and of what other backends
generate (e.g., C99).

A prior commit has updated the internals of `copilot-bluespec` to invert the
order in which Verilog inputs and outputs are declared. This commit updates the
prose in the `DESIGN.md` document to reflect the new design. The new design is
just different enough from the old one that it made sense to completely rewrite
certain portions of the document. One of the more notable changes is that there
are now separate sections about running the generated Verilog code on hardware
versus simulating the generated code in software, as these are distinct use
cases that require separate discussion.