copilot-bluespec: Fix translation of constant floating-point values. Refs #696.
copilot-bluespec generates syntactically invalid Bluespec code for special
floating-point values, such as negative zero, infinity, and NaN values. For
instance, it compiles infinity values to Infinity, which is not a valid
Bluespec floating-point value.

This patch modifies the function (`constFP`) that translates constant floating
point values from Copilot to Bluespec so that it correctly handles special
cases such as infs, NaNs, and -0.0.